project_alva/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
click==7.1.2
cycler==0.10.0
Django==3.1.2
django-bootstrap3==14.2.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-tables2==2.3.3
flake8==3.8.4
importlib-metadata==1.7.0
kiwisolver==1.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.1
numpy==1.19.4
pandas==1.1.4
pathspec==0.8.0
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pygal==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.28
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
xgboost==0.90
zipp==3.4.0
